rename: rename 'metadata-decorators' folder -> 'decorators'
