+++ src/decorators/do-not-throw-on-empty.ts
